+++ compose-miuix/shared/src/iosMain/kotlin/com/perol/pixez/shared/platform/IllustClipboard.ios.kt
@@ -9,4 +9,8 @@ actual class IllustClipboard {
     actual fun copy(text: String) {
         UIPasteboard.generalPasteboard.string = text
     }
+
+    actual fun getText(): String? {
+        return UIPasteboard.generalPasteboard.string
+    }
 }